Remove non-state Graph
- Not used in any examples/docs
