+++ libs/langgraph/langgraph/graph/__init__.py
@@ -1,11 +1,10 @@
-from langgraph.graph.graph import END, START, Graph
+from langgraph.constants import END, START
 from langgraph.graph.message import MessageGraph, MessagesState, add_messages
 from langgraph.graph.state import StateGraph
 
 __all__ = [
     "END",
     "START",
-    "Graph",
     "StateGraph",
     "MessageGraph",
     "add_messages",
