+++ libs/langgraph/langgraph/utils/pydantic.py
@@ -46,7 +46,7 @@ def get_fields(
         return model.model_fields
 
     if hasattr(model, "__fields__"):
-        return model.__fields__  # type: ignore[return-value]
+        return model.__fields__
     msg = f"Expected a Pydantic model. Got {type(model)}"
     raise TypeError(msg)
